登入身份驗證、Token驗證與授權管理、訂單、自動映射、授權

+++ Product.Core/Dto/CustomerBasketDto.cs
@@ -9,12 +9,13 @@ namespace Product.Core.Dto
 {
     public class CustomerBasketDto
     {
+
         [Required]
         public string Id { get; set; }
         public List<BasketItemDto> BasketItems { get; set; } = new List<BasketItemDto>();
         public int? DeliveryMethodId { get; set; }
         public string ClientSecret { get; set; }
         public string PaymentIntentId { get; set; }
-        public decimal ShippingPrice { get; set; }
+        public int ShippingPrice { get; set; }
     }
 }
